requirements.txt
# Requirements para PGD RAG
# Projeto de processamento de PDFs com extração de texto e análise visual

# Processamento de PDFs
PyMuPDF==1.26.4

# Google Gemini Vision API
google-generativeai==0.8.5

# Processamento de imagens
pillow==11.3.0

# Pinecone vector database
pinecone==7.3.0

# Carregamento de variáveis de ambiente
python-dotenv==1.1.1

# Barras de progresso
tqdm==4.67.1
